Cleaned up unused code
Moved writing of git config to cloning logic. No point in writing it
over and over...

Removed preference keys for remote name and branch. The default branch
is set on the remote, and will always be used; we don't need to know the
name. We also never switch branches with this solution - we just push
our current HEAD to a designated remote branch.