Issue #613: Add a "repoRelativePath" attribute to the Rook class
This was prompted by two facts:

1. The URIs of DocumentRepo files (content:// URIs) are opaque and
cannot be parsed to find out the file's location relative to the
repository root directory. I previously parsed them based on their
appearance in my virtual Android devices, but they may look very
different when using other DocumentsProviders, i.e. other backing
storage, such as a WebDAV mount.
2. When we added support for repository subfolders, this
repository-relative file location became very important.

This proposed change means that the "repo-relative path" becomes our
primary means of linking a book to a repository file. Before this, we
relied much more on the file's URI, but this is now less important,
although it is still necessary when dealing with DocumentRepo files,
i.e. the local directory repository type.

One problem with this solution is that there is now a code path (in
com.orgzly.android.BookName) for Rooks in the DB which do not yet have
a repoRelativePath value, and this code for the DocumentRepo (content://
directories) is quite complex, since it walks through the file tree
to find a matching file object. It will be very difficult to know how
long we need to maintain this code, so it would be worth a lot to find
a solution which lets us remove it entirely. But we do need some way to
deduce a repo-relative path for those existing remote books which do not
already have one stored in DB.