fix(deps): update versions.jgit to v6.6.1
I cleaned up the dependencies a bit. net.i2p.crypto:eddsa is no longer a
dependency of JGit, but we use those classes when handling Ed25519 keys.
We should update that code to get rid of the dependency, but I got lost
the last time I tried.

The current latest version of JGit is v7.3.0. See
https://bugs.eclipse.org/bugs/show_bug.cgi?id=582423#c0 for the reason
why I chose v6.6.1, instead of anything higher. I experienced the
exact same error as in the linked discussion on API 29 with v7.3.0.
The readNBytes() method was added in API 33, so using JGit > 6.6 would
require us to remove Git repository support on devices below API 33
(Android 13). We can do that some day, perhaps not too far in the
future, but I feel it's a bit too early.