Try adding a isDisplayed() check in onNoteInBook()
Line 194 in onRecyclerViewItem() currently seems to be the most common
place where flakiness hits. The fragment_book_recycler_view is not yet
present in the view hierarchy right after opening a book.

Let's try adding this kind of isDisplayed() check first, to see if that
helps.

Another thing worth looking into may be the allOf(..., isDisplayed())
pattern, which is used in some other places, seemingly with success?